Update JavaDoc for Zoom plugin
